Refactor app.py into Flask package structure
